feat: HapticManager 추가 및 적용
