+++ Capple/Capple/Presentation/Helper/PopGestureDisabledViewModifier.swift
@@ -7,25 +7,6 @@
 
 import SwiftUI
 
-final class PopGestureManager {
-    static let shared = PopGestureManager()
-    private init() {}
-    
-    /// Pop 제스처를 허용하는지 확인 변수
-    private var isAllowPopGestureValue = true
-    
-    /// Pop 제스처 받아오기
-    func isAllowPopGesture() -> Bool {
-        return isAllowPopGestureValue
-    }
-    
-    /// Pop 제스처를 허용하는 변수 업데이트
-    @MainActor
-    func updateAllowPopGesture(_ bool: Bool) async {
-        isAllowPopGestureValue = bool
-    }
-}
-
 struct PopGestureDisabledViewModifier: ViewModifier {
     
     func body(content: Content) -> some View {
