+++ Capple/Capple/Presentation/SearchResultView/View/SearchResultView.swift
@@ -19,16 +19,18 @@ struct SearchResultView: View {
                     principalView: {
                         HStack(spacing: 28) {
                             Button {
+                                HapticManager.shared.impact(style: .medium)
                                 tab = .answering
                             } label: {
-                                Text("오늘의 질문")
+                                Text("오늘의질문")
                                     .font(.pretendard(.semiBold, size: 14))
                                     .foregroundStyle(TextLabel.sub4)
                             }
                             Button {
+                                HapticManager.shared.impact(style: .medium)
                                 tab = .collecting
                             } label: {
-                                Text("질문 리스트")
+                                Text("질문리스트")
                                     .font(.pretendard(.semiBold, size: 14))
                                     .foregroundStyle(TextLabel.main)
                             }
